matrix * vector

+++ srcs/matrix/print_to.cpp
@@ -1,10 +1,10 @@
 #include "matrix.hpp"
 
 void Matrix::print_to(std::ostream &os) const {
-	for (int i = 0; i < rows_; i++)
+	for (size_t i = 0; i < rows_; i++)
     {
         os << "[";
-        for (int j = 0; j < cols_; j++)
+        for (size_t j = 0; j < cols_; j++)
         {
             os << data_[i][j];
             if (j < cols_ - 1)
